refactor: 구매 금액 입력과 구매한 로또 view 클래스 분리
